Cambio en categoria y sedes para que no quede selecionado el menu. Refactor de funioes de mailing.

+++ model/equipos.php
@@ -381,10 +381,15 @@ class Equipos {
 		$db->close();
 	}
 	
-	function seEnvioCorreoReserva($idEquipo= "", $idFecha="") {
+	function seEnvioCorreo($idEquipo= "", $idFecha="", $tabla="") {
 		$db = new Db();
 		
-		$query = "Select count(*) as cantidad from ga_correo_reservas where id_equipo = '$idEquipo' and id_fecha = '$idFecha'";
+		if ($tabla == "r") {
+			$query = "Select count(*) as cantidad from ga_correo_reservas where id_equipo = '$idEquipo' and id_fecha = '$idFecha'";
+		}
+		if ($tabla == "c") {
+			$query = "Select count(*) as cantidad from ga_correo_confirmacion where id_equipo = '$idEquipo' and id_fecha = '$idFecha'";
+		}
 
 		$res = $db->getRow($query); 
 	
@@ -397,64 +402,34 @@ class Equipos {
 		}
 	}
 	
-	function cargarCorreoReserva($idEquipo= "", $idFecha="") {
+	function cargarCorreo($idEquipo= "", $idFecha="", $tabla="") {
 		$db = new Db();
 
 		$today = date('Y-m-d');
 		
-		$query = "Insert into ga_correo_reservas(id_equipo,id_fecha,fecha_envio) value ($idEquipo,$idFecha,'$today')";
-		
-		$db->query($query);
-	
-		$db->close();
-	}
-	
-	function eliminarCorreoReserva($idEquipo= "", $idFecha="") {
-		$db = new Db();
-
-		$today = date('Y-m-d');
-		
-		$query = "delete from ga_correo_reservas where id_equipo = $idEquipo and id_fecha = $idFecha";
-		
-		$db->query($query);
-	
-		$db->close();
-	}
-	
-	function seEnvioCorreoConfirmacion($idEquipo= "", $idFecha="") {
-		$db = new Db();
-		
-		$query = "Select count(*) as cantidad from ga_correo_confirmacion where id_equipo = '$idEquipo' and id_fecha = '$idFecha'";
-
-		$res = $db->getRow($query); 
-	
-		$db->close();
-		
-		if($res->cantidad == 0) {
-			return false;
-		} else {
-			return true;
+		if ($tabla == "r") {
+			$query = "Insert into ga_correo_reservas(id_equipo,id_fecha,fecha_envio) value ($idEquipo,$idFecha,'$today')";
 		}
-	}
-	
-	function cargarCorreoConfirmacion($idEquipo= "", $idFecha="") {
-		$db = new Db();
-
-		$today = date('Y-m-d');
-		
-		$query = "Insert into ga_correo_confirmacion(id_equipo,id_fecha,fecha_envio) value ($idEquipo,$idFecha,'$today')";
+		if ($tabla == "c") {
+			$query = "Insert into ga_correo_confirmacion(id_equipo,id_fecha,fecha_envio) value ($idEquipo,$idFecha,'$today')";
+		}
 		
 		$db->query($query);
 	
 		$db->close();
 	}
 	
-	function eliminarCorreoConfirmacion($idEquipo= "", $idFecha="") {
+	function eliminarCorreo($idEquipo= "", $idFecha="", $tabla="") {
 		$db = new Db();
 
 		$today = date('Y-m-d');
 		
-		$query = "delete from ga_correo_confirmacion where id_equipo = $idEquipo and id_fecha = $idFecha";
+		if ($tabla == "r") {
+			$query = "delete from ga_correo_reservas where id_equipo = $idEquipo and id_fecha = $idFecha";
+		}
+		if ($tabla == "c") {
+			$query = "delete from ga_correo_confirmacion where id_equipo = $idEquipo and id_fecha = $idFecha";
+		}
 		
 		$db->query($query);
